G4-126-Fix-User-Swagger-Test(GET) GET/api/users/role/{role}Getting all Users by Role,GET/api/users/manager/orders,  GET/api/users/confectioner/orders,  GET/api/users/client/orders
